git commit -m "feat: Implementación completa de la gestión de transacciones" -m "Este commit introduce la funcionalidad completa para gestionar transacciones, incluyendo:
- Creación, edición y eliminación de transacciones.
- Gestión de categorías personalizadas.
- Asignación automática del tipo de transacción (ingreso/gasto) al seleccionar una categoría.
- Visualización de la categoría en la lista principal de transacciones.
- Correcciones de errores y mejoras de la interfaz."

+++ app/src/main/java/com/jmdevs/agendapersonal/data/local/entity/Transaction.kt
@@ -7,13 +7,13 @@ import java.util.Date
 @Entity(tableName = "transactions")
 data class Transaction(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val amount: Double,
+    val id: Long = 0,
+    val amount: Long,
     val type: String, // "INGRESO" o "GASTO"
     val category: String,
     val description: String,
     val date: Long,
     val paymentMethod: String,
     val currency: String = "Clp",
-    val balance: Double
+    val balance: Long
 )
